@@ -10,11 +10,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=7.1",
         "ocramius/lazy-map": "~1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "7.0.2"
     },
     "autoload": {
         "psr-0": {"lf4php": "src/"}
