@@ -1,35 +1,15 @@
 <?php
-/*
- * Copyright (c) 2012 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php;
 
 use Exception;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @author Janos Szurovecz <szjani@szjani.hu>
  */
-class LocationLoggerTest extends PHPUnit_Framework_TestCase
+class LocationLoggerTest extends TestCase
 {
     private $logger;
 
@@ -61,7 +41,7 @@ class LocationLoggerMock extends LocationLogger
 
     }
 
-    public function getName()
+    public function getName() : string
     {
 
     }
@@ -71,27 +51,27 @@ class LocationLoggerMock extends LocationLogger
         return $this->getShortLocation();
     }
 
-    public function isDebugEnabled()
+    public function isDebugEnabled() : bool
     {
 
     }
 
-    public function isErrorEnabled()
+    public function isErrorEnabled() : bool
     {
 
     }
 
-    public function isInfoEnabled()
+    public function isInfoEnabled() : bool
     {
 
     }
 
-    public function isTraceEnabled()
+    public function isTraceEnabled() : bool
     {
 
     }
 
-    public function isWarnEnabled()
+    public function isWarnEnabled() : bool
     {
 
     }
