@@ -1,39 +1,20 @@
 <?php
-/*
- * Copyright (c) 2012 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php;
 
-use PHPUnit_Framework_TestCase;
+use lf4php\nop\NOPLoggerFactory;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @author Janos Szurovecz <szjani@szjani.hu>
  */
-class LoggerFactoryTest extends PHPUnit_Framework_TestCase
+class LoggerFactoryTest extends TestCase
 {
     public function testFindKnownBindingsMissing()
     {
         $factory = LoggerFactory::getILoggerFactory();
-        self::assertInstanceOf('\lf4php\nop\NOPLoggerFactory', $factory);
+        self::assertInstanceOf(NOPLoggerFactory::class, $factory);
     }
 
     /**
@@ -41,7 +22,7 @@ class LoggerFactoryTest extends PHPUnit_Framework_TestCase
      */
     public function shouldUseFoundILoggerFactory()
     {
-        $factory = $this->getMock('\lf4php\ILoggerFactory');
+        $factory = $this->getMockBuilder(ILoggerFactory::class)->getMock();
         $loggerName = 'logger-name';
         LoggerFactoryChild::setILoggerFactory($factory);
         $factory
