@@ -1,25 +1,5 @@
 <?php
-/*
- * Copyright (c) 2012-2014 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php;
 
@@ -37,7 +17,7 @@ final class InitializedFactory extends RegistrationProvidedLoggerFactoryProxy
      * @param Logger $logger
      * @throws \RuntimeException
      */
-    public function registerLogger($classOrNamespace, Logger $logger)
+    public function registerLogger(string $classOrNamespace, Logger $logger) : void
     {
         $this->throwException();
     }
@@ -46,12 +26,12 @@ final class InitializedFactory extends RegistrationProvidedLoggerFactoryProxy
      * @param Logger $logger
      * @throws \RuntimeException
      */
-    public function setRootLogger(Logger $logger)
+    public function setRootLogger(Logger $logger) : void
     {
         $this->throwException();
     }
 
-    private function throwException()
+    private function throwException() : void
     {
         throw new RuntimeException(
             "Cannot register any Logger instances after the first call of getLogger() method"
