@@ -1,25 +1,5 @@
 <?php
-/*
- * Copyright (c) 2014 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php;
 
@@ -42,7 +22,7 @@ class LoggerMap extends AbstractLazyMap implements RegistrationProvidedLoggerFac
     /**
      * @return Logger
      */
-    public function getRootLogger()
+    public function getRootLogger() : Logger
     {
         return $this->rootLogger;
     }
@@ -50,7 +30,7 @@ class LoggerMap extends AbstractLazyMap implements RegistrationProvidedLoggerFac
     /**
      * @param Logger $rootLogger
      */
-    public function setRootLogger(Logger $rootLogger)
+    public function setRootLogger(Logger $rootLogger) : void
     {
         $this->rootLogger = $rootLogger;
     }
@@ -59,7 +39,7 @@ class LoggerMap extends AbstractLazyMap implements RegistrationProvidedLoggerFac
      * @param string $name
      * @return Logger
      */
-    public function getLogger($name)
+    public function getLogger(string $name) : Logger
     {
         $key = (string) $name;
         return $this->$key;
@@ -69,7 +49,7 @@ class LoggerMap extends AbstractLazyMap implements RegistrationProvidedLoggerFac
      * @param string $classOrNamespace
      * @param Logger $logger
      */
-    public function registerLogger($classOrNamespace, Logger $logger)
+    public function registerLogger(string $classOrNamespace, Logger $logger) : void
     {
         $key = $this->trim((string) $classOrNamespace);
         $this->$key = $logger;
@@ -94,7 +74,7 @@ class LoggerMap extends AbstractLazyMap implements RegistrationProvidedLoggerFac
         }
     }
 
-    private function trim($name)
+    private function trim(string $name) : string
     {
         return trim($name, self::SEPARATOR);
     }
