@@ -1,25 +1,5 @@
 <?php
-/*
- * Copyright (c) 2012 Janos Szurovecz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
+declare(strict_types=1);
 
 namespace lf4php\nop;
 
@@ -39,7 +19,7 @@ class NOPLogger extends NamedLoggerBase
     /**
      * @return NOPLogger
      */
-    public static function getNOPLogger()
+    public static function getNOPLogger() : NOPLogger
     {
         if (self::$instance === null) {
             self::$instance = new self();
@@ -72,27 +52,27 @@ class NOPLogger extends NamedLoggerBase
     {
     }
 
-    public function isDebugEnabled()
+    public function isDebugEnabled() : bool
     {
         return false;
     }
 
-    public function isErrorEnabled()
+    public function isErrorEnabled() : bool
     {
         return false;
     }
 
-    public function isInfoEnabled()
+    public function isInfoEnabled() : bool
     {
         return false;
     }
 
-    public function isTraceEnabled()
+    public function isTraceEnabled() : bool
     {
         return false;
     }
 
-    public function isWarnEnabled()
+    public function isWarnEnabled() : bool
     {
         return false;
     }
